Add files via upload
This contains the .hfe CP/M file for the Kaypro II with the 81-149 ROM. These file can be put on a USB stick for a GoTek floppy emulator and directly be used. 
The "FF.CFG" and "IMAGE_A.CFG" files need to be also on the USB stick if used for a GoTek. 

The files are tested on a Kaypro II with a GoTek installed with FlashFloppy.